Refactor plotting engine into dedicated modules
