+++ engine/__init__.py
@@ -1,0 +1,24 @@
+from .config import infer_parameters, normalize_layout, normalize_plots
+from .data_pipeline import apply_preprocessing, prepare_datafile
+from .runner import process_plot, run_batch, run_gnuplot_script
+from .script_builder import (
+    compute_residual_metrics,
+    estimate_initial_params,
+    generate_gnuplot_code,
+    parse_fit_output,
+)
+
+__all__ = [
+    "apply_preprocessing",
+    "compute_residual_metrics",
+    "estimate_initial_params",
+    "generate_gnuplot_code",
+    "infer_parameters",
+    "normalize_layout",
+    "normalize_plots",
+    "parse_fit_output",
+    "prepare_datafile",
+    "process_plot",
+    "run_batch",
+    "run_gnuplot_script",
+]
